requirements.txt
pyserial==3.5
dmx485==1.2.1
mido==1.3.0